refactor(adapters): extract MCP loader + cache-control + wrapper script from claude.py (audit-142)
